requirements.txt
groq==0.3.3
rich==13.5.2
python-dotenv==1.0.0